Monotone-Parent: da1af898a9c7e2aa99b063087444d269a67e52c5 Monotone-Revision: 4cdfac2b03944c4ddb408b6d5e6618620cdf334f
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2010-08-30T21:42:26
Monotone-Branch: ca.inverse.sope

+++ sope-mime/NGMime/NGMimeRFC822DateHeaderFieldParser.m
@@ -145,7 +145,7 @@ static NSTimeZone *parseTimeZone(const char *s, unsigned int len) {
   char *newString, *digits;
   const char *p;
   NSTimeZone *tz;
-  unsigned int hours, minutes, seconds, remaining;
+  NSInteger hours, minutes, seconds, remaining;
   int sign;
 
   sign = 1;
